requirements.txt
git+https://github.com/huggingface/transformers
datasets 
loralib
sentencepiece
sentence_transformers
langchain
llama-index